+++ src/main/java/unicash/logic/commands/FindCommand.java
@@ -6,7 +6,7 @@ import unicash.commons.enums.CommandType;
 import unicash.commons.util.ToStringBuilder;
 import unicash.logic.UniCashMessages;
 import unicash.model.Model;
-import unicash.model.transaction.predicates.TransactionContainsAnyKeywordsPredicate;
+import unicash.model.transaction.predicates.TransactionContainsAllKeywordsPredicate;
 
 /**
  * Finds and lists all transactions in UniCa$h whose name contains any of the argument keywords.
@@ -17,13 +17,13 @@ public class FindCommand extends Command {
     public static final String COMMAND_WORD = CommandType.FIND.getCommandWords();
     public static final String MESSAGE_USAGE = CommandType.FIND.getMessageUsage();
 
-    private final TransactionContainsAnyKeywordsPredicate predicate;
+    private final TransactionContainsAllKeywordsPredicate predicate;
 
     /**
      * Creates a {@code TransactionContainsAnyKeywordsPredicate} with a non-null
      * predicate.
      */
-    public FindCommand(TransactionContainsAnyKeywordsPredicate predicate) {
+    public FindCommand(TransactionContainsAllKeywordsPredicate predicate) {
         requireNonNull(predicate);
         this.predicate = predicate;
     }
