+++ src/main/java/unicash/logic/parser/UniCashParser.java
@@ -96,9 +96,6 @@ public class UniCashParser {
         case SUMMARY:
             return new SummaryCommand();
 
-        //case FilterCommand.COMMAND_WORD:
-        //    return new FilterCommandParser().parse(arguments);
-
         default:
             logger.finer("This user input caused a ParseException: " + userInput);
             throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
