+++ src/test/java/unicash/logic/parser/FindCommandParserTest.java
@@ -1,18 +1,35 @@
 package unicash.logic.parser;
 
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static unicash.logic.UniCashMessages.MESSAGE_INVALID_COMMAND_FORMAT;
-
-import java.util.Arrays;
+import static unicash.logic.commands.CommandTestUtil.AMOUNT_DESC_NUS;
+import static unicash.logic.commands.CommandTestUtil.DATETIME_DESC_NUS;
+import static unicash.logic.commands.CommandTestUtil.INVALID_AMOUNT_DESC;
+import static unicash.logic.commands.CommandTestUtil.INVALID_DATETIME_DESC;
+import static unicash.logic.commands.CommandTestUtil.INVALID_LOCATION_DESC;
+import static unicash.logic.commands.CommandTestUtil.INVALID_TRANSACTION_NAME_DESC;
+import static unicash.logic.commands.CommandTestUtil.INVALID_TYPE_DESC;
+import static unicash.logic.commands.CommandTestUtil.TRANSACTION_NAME_DESC_NUS;
+import static unicash.logic.commands.CommandTestUtil.TYPE_DESC_EXPENSE;
+import static unicash.logic.commands.CommandTestUtil.TYPE_DESC_INCOME;
+import static unicash.logic.parser.CliSyntax.PREFIX_TYPE;
+import static unicash.logic.parser.CommandParserTestUtil.assertParseFailure;
+import static unicash.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
 import unicash.commons.util.ToStringBuilder;
-import unicash.logic.commands.FindCommand;
-import unicash.model.transaction.predicates.TransactionContainsAnyKeywordsPredicate;
+import unicash.logic.UniCashMessages;
+import unicash.logic.parser.exceptions.ParseException;
+import unicash.model.commons.Amount;
+import unicash.model.transaction.DateTime;
+import unicash.model.transaction.Location;
+import unicash.model.transaction.Type;
+import unicash.model.transaction.predicates.TransactionContainsAllKeywordsPredicate;
+
 
 /**
  * A class to test the FindCommandParser.
@@ -23,24 +40,40 @@ public class FindCommandParserTest {
 
     @Test
     public void parse_emptyArg_throwsParseException() {
-        CommandParserTestUtil.assertParseFailure(parser, "     ",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
-        CommandParserTestUtil.assertParseFailure(parser, "",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
+        String emptyArgument = "";
+        assertThrows(ParseException.class, () -> parser.parse(emptyArgument));
+
+    }
+
+    @Test
+    public void parse_repeatedType_failure() {
+        assertParseFailure(parser, TYPE_DESC_INCOME + TYPE_DESC_EXPENSE,
+                UniCashMessages.getErrorMessageForDuplicatePrefixes(PREFIX_TYPE));
     }
 
 
     @Test
-    public void parse_validArgs_returnsFindCommand() {
-        // no leading and trailing whitespaces
-        FindCommand expectedFindCommand =
-                new FindCommand(new TransactionContainsAnyKeywordsPredicate(Arrays.asList("Shopping", "Work")));
-        CommandParserTestUtil.assertParseSuccess(parser, "Shopping Work", expectedFindCommand);
+    public void parse_invalidValue_failure() {
+        // invalid name
+        assertParseFailure(parser, INVALID_TRANSACTION_NAME_DESC + AMOUNT_DESC_NUS + DATETIME_DESC_NUS
+                + TYPE_DESC_EXPENSE, unicash.model.transaction.Name.MESSAGE_CONSTRAINTS);
 
-        // multiple whitespaces between keywords
-        CommandParserTestUtil.assertParseSuccess(parser, " \n Shopping \n \t Work  \t", expectedFindCommand);
+        // invalid amount
+        assertParseFailure(parser, TRANSACTION_NAME_DESC_NUS + INVALID_AMOUNT_DESC
+                + DATETIME_DESC_NUS + TYPE_DESC_EXPENSE, Amount.MESSAGE_CONSTRAINTS);
+
+        // invalid type
+        assertParseFailure(parser, TRANSACTION_NAME_DESC_NUS + AMOUNT_DESC_NUS + DATETIME_DESC_NUS
+                + INVALID_TYPE_DESC, Type.MESSAGE_CONSTRAINTS);
+
+        // invalid location
+        assertParseFailure(parser, TRANSACTION_NAME_DESC_NUS + AMOUNT_DESC_NUS + DATETIME_DESC_NUS
+                + TYPE_DESC_EXPENSE + INVALID_LOCATION_DESC, Location.MESSAGE_CONSTRAINTS);
+
+        // invalid datetime
+        assertParseFailure(parser, TRANSACTION_NAME_DESC_NUS + AMOUNT_DESC_NUS + INVALID_DATETIME_DESC
+                + TYPE_DESC_EXPENSE, DateTime.MESSAGE_CONSTRAINTS);
     }
-
 
     @Test
     public void sameFindCommandParser_equalsTrue() {
@@ -66,9 +99,12 @@ public class FindCommandParserTest {
     @Test
     public void toStringTest() {
         FindCommandParser findCommandParser = new FindCommandParser();
-        String expected = new ToStringBuilder(new FindCommandParser()).toString();
+        TransactionContainsAllKeywordsPredicate filterPredicate =
+                new TransactionContainsAllKeywordsPredicate();
+
+        String expected = new ToStringBuilder(new FindCommandParser())
+                .add("filterPredicate", filterPredicate).toString();
         assertEquals(expected, findCommandParser.toString());
     }
 
 }
-
